fix: flyway 타입 수정

+++ backend/src/main/resources/db/migration/V8__change_column_tier.sql
@@ -1,1 +1,2 @@
-alter table member change column tier experience bigint not null default 0;
+alter table member
+    change column tier experience int not null default 0;
